requirements.txt
selenium==4.15.2
aiogram==3.4.1
python-dotenv==1.0.0